requirements.txt
alpaca-py
pandas
matplotlib
numpy
scikit-learn
python-dotenv
yfinance
lxml